+++ src/components/datagrid/datagrid.js
@@ -104,6 +104,7 @@ const COMPONENT_NAME = 'datagrid';
  * @param {Function} [settings.onDestroyCell=null] A call back that goes along with onPostRenderCel and will fire when this cell is destroyed and you need noification of that.
  * @param {Function} [settings.onEditCell=null] A callback that fires when a cell is edited, the editor object is passed in to the function
  * @param {Function} [settings.onExpandRow=null] A callback function that fires when expanding rows. To be used. when expandableRow is true. The function gets eventData about the row and grid and a response function callback. Call the response function with markup to append and delay opening the row.
+ * @param {Function} [settings.onKeyDown=null] A callback function that fires when any key is pressed down.
  * @param {boolean}  [settings.searchExpandableRow=true] If true keywordSearch will search in expandable rows (default). If false it will not search expandable rows.
  * @param {object}   [settings.emptyMessage]
  * @param {object}   [settings.emptyMessage.title='No Data Available']
@@ -197,6 +198,7 @@ const DATAGRID_DEFAULTS = {
   onDestroyCell: null,
   onEditCell: null,
   onExpandRow: null,
+  onKeyDown: null,
   emptyMessage: { title: (Locale ? Locale.translate('NoData') : 'No Data Available'), info: '', icon: 'icon-empty-no-data' },
   searchExpandableRow: true,
   allowChildExpandOnMatch: false
@@ -441,6 +443,7 @@ Datagrid.prototype = {
     }
     // Add row status
     data.rowStatus = { icon: 'new', text: Locale.translate('New'), tooltip: Locale.translate('New') };
+    this.saveDirtyRows();
 
     // Add to array
     const appendArray = this.settings.groupable &&
@@ -452,6 +455,7 @@ Datagrid.prototype = {
       appendArray.splice(location, 0, data);
     }
 
+    this.restoreDirtyRows();
     this.setRowGrouping();
     this.pagerRefresh(location);
     this.syncSelectedRowsIdx();
@@ -7779,7 +7783,6 @@ Datagrid.prototype = {
         }
         return self.dataRowIndex(visibleRow);
       };
-
       if (!node.length) {
         self.activeCell.node = self.cellNode(row, cell);
         node = self.activeCell.node;
@@ -7817,6 +7820,23 @@ Datagrid.prototype = {
       const lastRow = visibleRows.last();
       const lastCell = self.settings.columns.length - 1;
 
+      if (self.settings.onKeyDown) {
+        const response = (isCancelled) => {
+          if (!isCancelled) {
+            e.stopPropagation();
+            e.preventDefault();
+          }
+        };
+
+        const args = { activeCell: self.activeCell, row, cell };
+        const ret = self.settings.onKeyDown(e, args, response);
+        if (ret === false || !response) {
+          e.stopPropagation();
+          e.preventDefault();
+          return;
+        }
+      }
+
       // Tab, Left, Up, Right and Down arrow keys.
       if ([9, 37, 38, 39, 40].indexOf(key) !== -1) {
         if (target.closest('.code-block').length &&
@@ -7835,15 +7855,6 @@ Datagrid.prototype = {
 
       // Tab, Left and Right arrow keys.
       if ([9, 37, 39].indexOf(key) !== -1) {
-        if (key === 9 && self.settings.onKeyDown) {
-          const ret = self.settings.onKeyDown(e);
-          if (ret === false) {
-            e.stopPropagation();
-            e.preventDefault();
-            return;
-          }
-        }
-
         if (key === 9 && self.editor && self.editor.name === 'input' && col.inlineEditor === true) {
           // Editor.destroy
           self.editor.destroy();
@@ -9327,6 +9338,12 @@ Datagrid.prototype = {
   setDirtyCell(row, cell, dirtyOptions) {
     const cellNode = this.cellNode(row, cell);
 
+    // Do not show dirty indicator on new cells or cells with errors on them
+    if (this.settings.dataset[row] && this.settings.dataset[row].rowStatus &&
+      (this.settings.dataset[row].rowStatus.icon === 'new' || row === 0)) {
+      return;
+    }
+
     if (dirtyOptions) {
       this.addToDirtyArray(row, cell, dirtyOptions);
     }
@@ -10109,10 +10126,10 @@ Datagrid.prototype = {
     }
     const sort = this.sortFunction(this.sortColumn.sortId, this.sortColumn.sortAsc);
 
-    this.setDirtyBeforeSort();
+    this.saveDirtyRows();
     this.settings.dataset.sort(sort);
     this.setTreeDepth();
-    this.setDirtyAfterSort();
+    this.restoreDirtyRows();
 
     // Resync the _selectedRows array
     if (this.settings.selectable) {
@@ -10121,11 +10138,9 @@ Datagrid.prototype = {
   },
 
   /**
-  * Set current data to sync up dirtyArray before sort
-  * @private
-  * @returns {void}
-  */
-  setDirtyBeforeSort() {
+   * Set current data to sync up dirtyArray before sort
+   */
+  saveDirtyRows() {
     const s = this.settings;
     const dataset = s.treeGrid ? s.treeDepth : s.dataset;
     if (s.showDirty && !this.settings.source && this.dirtyArray && this.dirtyArray.length) {
@@ -10143,7 +10158,7 @@ Datagrid.prototype = {
   * @private
   * @returns {void}
   */
-  setDirtyAfterSort() {
+  restoreDirtyRows() {
     const s = this.settings;
     const dataset = s.treeGrid ? s.treeDepth : s.dataset;
     if (s.showDirty && this.dirtyArray && this.dirtyArray.length) {