+++ test/components/datagrid/datagrid.e2e-spec.js
@@ -264,12 +264,12 @@ describe('Datagrid Expandable Row Tests', () => {
       expect(size.height).toEqual(0);
     });
     await button.click();
-    await browser.driver.sleep(config.sleep);
+    await browser.driver.sleep(config.sleepLonger);
     await detailRow.getSize().then((size) => {
       expect(size.height).toBeGreaterThan(150);
     });
     await button.click();
-    await browser.driver.sleep(config.sleep);
+    await browser.driver.sleep(config.sleepLonger);
     await detailRow.getSize().then((size) => {
       expect(size.height).toEqual(0);
     });
@@ -904,12 +904,14 @@ describe('Datagrid paging tests', () => {
       expect(await element(by.css('#datagrid .datagrid-header th:nth-child(2).is-sorted-desc')).isPresent()).toBeFalsy();
 
       await element(by.css('#datagrid .datagrid-header th:nth-child(2)')).click();
+      await browser.driver.sleep(config.sleep);
       await element(by.css('#datagrid .datagrid-header th:nth-child(2)')).click();
       await browser.driver.sleep(config.sleep);
 
       expect(await element(by.css('#datagrid .datagrid-header th:nth-child(2).is-sorted-desc')).isPresent()).toBeTruthy();
 
       await element(by.css('.pager-next a')).click();
+      await browser.driver.sleep(config.sleep);
       await element(by.css('.pager-prev a')).click();
       await browser.driver.sleep(config.sleep);
 
@@ -1244,6 +1246,33 @@ describe('Datagrid editor dropdown source tests', () => {
   });
 });
 
+describe('Datagrid onKeyDown Tests', () => {
+  beforeEach(async () => {
+    await utils.setPage('/components/datagrid/test-editable-onkeydown?layout=nofrills');
+
+    const datagridEl = await element(by.css('#datagrid tr:nth-child(1)'));
+    await browser.driver
+      .wait(protractor.ExpectedConditions.presenceOf(datagridEl), config.waitsFor);
+  });
+
+  it('Should not have errors', async () => {
+    await utils.checkForErrors();
+  });
+
+  it('Should show toast on keydown', async () => {
+    await element(by.css('#datagrid .datagrid-body tbody tr:nth-child(1) td:nth-child(1)')).click();
+
+    const editCellSelector = '.has-editor.is-editing input';
+    const inputEl = await element(by.css(editCellSelector));
+    await browser.driver.wait(protractor.ExpectedConditions.presenceOf(inputEl), config.waitsFor);
+    await element(by.css(editCellSelector)).sendKeys('j');
+    await element(by.css(editCellSelector)).sendKeys(protractor.Key.TAB);
+
+    expect(await element.all(by.css('#toast-container .toast-message')).first().getText()).toEqual('You hit j. Event has been vetoed, so nothing will happen.');
+    expect(await element.all(by.css('#toast-container .toast-message')).last().getText()).toEqual('You hit Tab. Event has been vetoed, so nothing will happen.');
+  });
+});
+
 describe('Datagrid Header Alignment With Ellipsis', () => {
   beforeEach(async () => {
     await utils.setPage('/components/datagrid/test-ellipsis-header-align?layout=nofrills');
@@ -1354,6 +1383,56 @@ describe('Datagrid Header Overlapping Sorting Indicator', () => {
       expect(await browser.protractorImageComparison.checkElement(containerEl, 'datagrid-header-align-overlapping-sort-indicator')).toEqual(0);
     });
   }
+});
+
+describe('Datagrid Dirty and New Row Indicator', () => {
+  beforeEach(async () => {
+    await utils.setPage('/components/datagrid/test-overlaps-row-cell-indicators?layout=nofrills');
+
+    const datagridEl = await element(by.css('#datagrid tr:nth-child(1)'));
+    await browser.driver
+      .wait(protractor.ExpectedConditions.presenceOf(datagridEl), config.waitsFor);
+  });
+
+  it('Should not have errors', async () => {
+    await utils.checkForErrors();
+  });
+
+  it('Should show a new line indicator', async () => {
+    await element(by.id('add-row-top')).click();
+
+    expect(await element(by.css('#datagrid .datagrid-body tbody tr:nth-child(1) td:nth-child(1).rowstatus-cell')).isPresent()).toBe(true);
+    await utils.checkForErrors();
+  });
+
+  it('should show a dirty indicator on existing rows', async () => {
+    await element(by.css('#datagrid .datagrid-body tbody tr:nth-child(1) td:nth-child(1)')).click();
+
+    const editCellSelector = '.has-editor.is-editing input';
+    const inputEl = await element(by.css(editCellSelector));
+    await browser.driver.wait(protractor.ExpectedConditions.presenceOf(inputEl), config.waitsFor);
+    await element(by.css(editCellSelector)).sendKeys('121');
+
+    expect(await element(by.css(editCellSelector)).getAttribute('value')).toEqual('121');
+    await element(by.css(editCellSelector)).sendKeys(protractor.Key.ENTER);
+
+    expect(await element(by.css('#datagrid .datagrid-body tbody tr:nth-child(1) td:nth-child(1).is-dirty-cell')).isPresent()).toBe(true);
+  });
+
+  it('should not show a dirty indicator on new rows', async () => {
+    await element(by.id('add-row-top')).click();
+    await element(by.css('#datagrid .datagrid-body tbody tr:nth-child(1) td:nth-child(1)')).click();
+
+    const editCellSelector = '.has-editor.is-editing input';
+    const inputEl = await element(by.css(editCellSelector));
+    await browser.driver.wait(protractor.ExpectedConditions.presenceOf(inputEl), config.waitsFor);
+    await element(by.css(editCellSelector)).sendKeys('121');
+
+    expect(await element(by.css(editCellSelector)).getAttribute('value')).toEqual('121');
+    await element(by.css(editCellSelector)).sendKeys(protractor.Key.ENTER);
+
+    expect(await element(by.css('#datagrid .datagrid-body tbody tr:nth-child(1) td:nth-child(1).is-dirty-cell')).isPresent()).toBe(false);
+  });
 });
 
 describe('Datagrid contextmenu tests', () => {
@@ -1983,7 +2062,7 @@ describe('Datagrid paging indeterminate single select tests', () => {
   beforeEach(async () => {
     await utils.setPage('/components/datagrid/test-paging-select-indeterminate-single?layout=nofrills');
 
-    const datagridEl = await element(by.css('#datagrid tbody tr:nth-child(1)'));
+    const datagridEl = await element(by.css('#datagrid tbody tr:nth-child(5)'));
     await browser.driver
       .wait(protractor.ExpectedConditions.presenceOf(datagridEl), config.waitsFor);
   });
@@ -2038,7 +2117,9 @@ describe('Datagrid paging serverside multi select tests', () => {
 
   it('Should be able to select and have selections clear when paging', async () => {
     await element(by.css('#datagrid .datagrid-body tbody tr:nth-child(1) td:nth-child(1)')).click();
+    await browser.driver.sleep(config.sleep);
     await element(by.css('#datagrid .datagrid-body tbody tr:nth-child(2) td:nth-child(1)')).click();
+    await browser.driver.sleep(config.sleep);
 
     expect(await element.all(by.css('.datagrid-row.is-selected')).count()).toEqual(2);
 
@@ -2123,7 +2204,9 @@ describe('Datagrid paging serverside single select tests', () => {
 
   it('Should be able to select and have selections clear when paging', async () => {
     await element(by.css('#datagrid .datagrid-body tbody tr:nth-child(1) td:nth-child(1)')).click();
+    await browser.driver.sleep(config.sleep);
     await element(by.css('#datagrid .datagrid-body tbody tr:nth-child(2) td:nth-child(1)')).click();
+    await browser.driver.sleep(config.sleep);
 
     expect(await element.all(by.css('.datagrid-row.is-selected')).count()).toEqual(1);
 
@@ -2551,7 +2634,7 @@ describe('Datagrid tooltip tests', () => {
 
   it('Should show tooltip on header text cut off with ellipsis', async () => {
     await browser.actions().mouseMove(element(by.css('.datagrid-header thead th[data-column-id="orderDate"] .datagrid-column-wrapper'))).perform();
-    await browser.driver.sleep(config.sleep);
+    await browser.driver.sleep(config.sleepLonger);
     await browser.driver
       .wait(protractor.ExpectedConditions.presenceOf(await element(by.css('.grid-tooltip'))), config.waitsFor);
     const tooltip = await element(by.css('.grid-tooltip'));